WIP #948 When updating/deleting credentials, also update the protocols that reference those credentials (using transactions)

+++ Ui/Service/DataSource/DAO/Dapper/DapperDataBase_CredentialVault.cs
@@ -5,6 +5,8 @@ using Dapper;
 using _1RM.Model.Protocol.Base;
 using NUlid;
 using Newtonsoft.Json;
+using System.Data;
+using System.Windows;
 
 // ReSharper disable InconsistentNaming
 
@@ -14,7 +16,9 @@ namespace _1RM.Service.DataSource.DAO.Dapper
     {
         private void InitPasswordVault()
         {
-            _dbConnection?.Execute(NormalizedSql(@$"
+            lock (this)
+            {
+                _dbConnection?.Execute(NormalizedSql(@$"
 CREATE TABLE IF NOT EXISTS `{TableCredential.TABLE_NAME}` (
     `{nameof(TableCredential.Id)}`       VARCHAR (64) PRIMARY KEY
                                               NOT NULL
@@ -26,6 +30,7 @@ CREATE TABLE IF NOT EXISTS `{TableCredential.TABLE_NAME}` (
     `{nameof(TableCredential.Json)}`     TEXT         NOT NULL
 );
 "));
+            }
         }
 
         public override ResultSelects<Credential> GetCredentials()
@@ -52,9 +57,9 @@ CREATE TABLE IF NOT EXISTS `{TableCredential.TABLE_NAME}` (
         }
 
         private string SqlInsertCredentialVault => NormalizedSql($@"INSERT INTO `{TableCredential.TABLE_NAME}`
-(`{nameof(TableCredential.Name)}`,`{nameof(TableCredential.Hash)}`, `{nameof(TableCredential.Json)}`)
+(`{nameof(TableCredential.Id)}`,`{nameof(TableCredential.Name)}`,`{nameof(TableCredential.Hash)}`, `{nameof(TableCredential.Json)}`)
 VALUES
-(@{nameof(TableCredential.Name)}, @{nameof(TableCredential.Hash)}, @{nameof(TableCredential.Json)});");
+(@{nameof(TableCredential.Id)}, @{nameof(TableCredential.Name)}, @{nameof(TableCredential.Hash)}, @{nameof(TableCredential.Json)});");
 
         public override Result AddCredential(ref Credential credential)
         {
@@ -94,57 +99,53 @@ VALUES
 `{nameof(TableCredential.Hash)}` = @{nameof(TableCredential.Hash)},
 `{nameof(TableCredential.Json)}` = @{nameof(TableCredential.Json)}
 WHERE `{nameof(TableCredential.Id)}`= @{nameof(TableCredential.Id)};");
-        public override Result UpdateCredential(Credential credential)
+        public override Result UpdateCredential(Credential credential, List<ProtocolBaseWithAddressPortUserPwd>? relatedProtocols = null)
         {
             string info = IoC.Translate("We can not update on database:");
             lock (this)
             {
                 var result = OpenConnection(info);
-                if (!result.IsSuccess) return result;
+                if (!result.IsSuccess || _dbConnection == null) return result;
                 try
                 {
-                    var tpv = new TableCredential()
+                    bool ret = false;
+                    using (var transaction = _dbConnection.BeginTransaction())
                     {
-                        Id = credential.DatabaseId,
-                        Name = credential.Name,
-                        Hash = credential.GetHash(),
-                        Json = JsonConvert.SerializeObject(credential),
-                    };
-                    var ret = _dbConnection?.Execute(SqlUpdatePasswordVault, tpv) > 0;
-                    if (ret)
-                    {
-                        return Result.Success();
-                    }
-                    return Result.Fail(info, DatabaseName, "No rows affected during update operation.");
-                }
-                catch (Exception e)
-                {
-                    return ResultString.Fail(info, DatabaseName, e.Message);
-                }
-            }
-        }
-
-        public override Result UpdateCredential(IEnumerable<Credential> credentials)
-        {
-            string info = IoC.Translate("We can not update on database:");
-            lock (this)
-            {
-                var result = OpenConnection(info);
-                if (!result.IsSuccess) return result;
-                try
-                {
-                    var items = new List<TableCredential>();
-                    foreach (var credential in credentials)
-                    {
+                        // make sure Address and Port are empty, because they are not used in credential vault
+                        credential.Address = "";
+                        credential.Port = "";
                         var tpv = new TableCredential()
                         {
+                            Id = credential.DatabaseId,
                             Name = credential.Name,
                             Hash = credential.GetHash(),
                             Json = JsonConvert.SerializeObject(credential),
                         };
-                        items.Add(tpv);
+                        ret = _dbConnection?.Execute(SqlUpdatePasswordVault, tpv) > 0;
+
+
+
+                        // after credential is updated, the related protocols will also be updated with the new credential information.
+                        if (ret && relatedProtocols?.Count > 0)
+                        {
+                            foreach (var protocol in relatedProtocols)
+                            {
+                                protocol.InheritedCredentialName = tpv.Name;
+                                protocol.SetCredential(credential);
+                            }
+                            ret = _dbConnection?.Execute(SqlUpdate, relatedProtocols.Select(x => x.ToDbServer())) > 0;
+                        }
+                        if (!ret)
+                        {
+                            transaction.Rollback();
+                        }
+                        else
+                        {
+                            transaction.Commit();
+                        }
                     }
-                    var ret = _dbConnection?.Execute(SqlUpdate, items) > 0;
+
+
                     if (ret)
                     {
                         return Result.Success();
@@ -158,20 +159,80 @@ WHERE `{nameof(TableCredential.Id)}`= @{nameof(TableCredential.Id)};");
             }
         }
 
-        public override Result DeleteCredential(IEnumerable<string> names)
+        //public override Result UpdateCredential(IEnumerable<Credential> credentials)
+        //{
+        //    string info = IoC.Translate("We can not update on database:");
+        //    lock (this)
+        //    {
+        //        var result = OpenConnection(info);
+        //        if (!result.IsSuccess) return result;
+        //        try
+        //        {
+        //            var items = new List<TableCredential>();
+        //            foreach (var credential in credentials)
+        //            {
+        //                var tpv = new TableCredential()
+        //                {
+        //                    Name = credential.Name,
+        //                    Hash = credential.GetHash(),
+        //                    Json = JsonConvert.SerializeObject(credential),
+        //                };
+        //                items.Add(tpv);
+        //            }
+        //            var ret = _dbConnection?.Execute(SqlUpdate, items) > 0;
+        //            if (ret)
+        //            {
+        //                return Result.Success();
+        //            }
+        //            return Result.Fail(info, DatabaseName, "No rows affected during update operation.");
+        //        }
+        //        catch (Exception e)
+        //        {
+        //            return ResultString.Fail(info, DatabaseName, e.Message);
+        //        }
+        //    }
+        //}
+
+        public override Result DeleteCredential(IEnumerable<string> names, List<ProtocolBaseWithAddressPortUserPwd>? relatedProtocols = null)
         {
             var info = IoC.Translate("We can not delete from database:");
+            if (!names.Any())
+                return Result.Success();
+
             lock (this)
             {
                 var result = OpenConnection(info);
-                if (!result.IsSuccess) return result;
+                if (!result.IsSuccess || _dbConnection == null) return result;
                 try
                 {
-                    // special case: dapper does not support IN operator for postgresql
-                    var sql = DatabaseType == DatabaseType.PostgreSQL
-                        ? $@"DELETE FROM `{TableCredential.TABLE_NAME}` WHERE `{nameof(TableCredential.Name)}` = ANY(@{nameof(TableCredential.Name)});"
-                        : $@"DELETE FROM `{TableCredential.TABLE_NAME}` WHERE `{nameof(TableCredential.Name)}` IN @{nameof(TableCredential.Name)};";
-                    var ret = _dbConnection?.Execute(NormalizedSql(sql), new { Name = names }) > 0;
+                    bool ret = false;
+                    using (var transaction = _dbConnection.BeginTransaction())
+                    {
+                        // special case: dapper does not support IN operator for postgresql
+                        var sql = DatabaseType == DatabaseType.PostgreSQL
+                            ? $@"DELETE FROM `{TableCredential.TABLE_NAME}` WHERE `{nameof(TableCredential.Name)}` = ANY(@{nameof(TableCredential.Name)});"
+                            : $@"DELETE FROM `{TableCredential.TABLE_NAME}` WHERE `{nameof(TableCredential.Name)}` IN @{nameof(TableCredential.Name)};";
+                        ret = _dbConnection?.Execute(NormalizedSql(sql), new { Name = names }) > 0;
+
+                        // after credential is deleted, the related protocols will also be updated to clear the credential information.
+                        if (ret && relatedProtocols?.Count > 0)
+                        {
+                            foreach (var protocol in relatedProtocols)
+                            {
+                                protocol.InheritedCredentialName = ""; // clear the credential name
+                            }
+                            ret = _dbConnection?.Execute(SqlUpdate, relatedProtocols.Select(x => x.ToDbServer())) > 0;
+                        }
+                        if (!ret)
+                        {
+                            transaction.Rollback();
+                        }
+                        else
+                        {
+                            transaction.Commit();
+                        }
+                    }
+
                     if (ret)
                         return Result.Success();
                     return Result.Fail(info, DatabaseName, "No rows affected during delete operation.");
